requirements.txt
easydict==1.9
kornia==0.5.0
numpy==1.19.5
opencv_python_headless==4.5.3.56
Pillow==8.4.0
PyYAML==6.0
tqdm==4.62.2
